pyproject.toml
[tool.poetry]
name = "peac"
version = "0.2.3"
description = "Prompt Engineering as Code"
authors = ["gx1 <18548727+giper45@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "peac"}]
include = ["peac/template.yaml"]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
# Dipendenze core - versioni stabili per Windows
typer = "^0.9.0"
click = "~8.1.7"
pyyaml = "^6.0"
requests = "^2.31.0"
rich = "^13.0.0"
validators = "^0.22.0"
markdown = "^3.5.0"

# GUI dependencies - versioni stabili
customtkinter = "^5.2.0"
flet = "^0.24.0"

# Rimuovi le dipendenze problematiche come g4f e pythonnet
# g4f = "^0.1.0"  # RIMOSSO - problematico su Windows
# pythonnet = "^3.0.0"  # RIMOSSO - richiede build tools
beautifulsoup4 = "^4.13.5"
python-docx = "^1.2.0"
pdfplumber = "^0.11.7"
openpyxl = "^3.1.0"
faiss-cpu = "^1.8.0"
sentence-transformers = "^3.2.0"
torch = "^2.1.0"
numpy = "^2.3.3"
notebook = "^7.4.5"
ipykernel = "^6.30.1"
pandas = "^2.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
bump-my-version = "^0.26.0"
# PyInstaller development version that supports Python 3.13
pyinstaller = {git = "https://github.com/pyinstaller/pyinstaller.git"}

[tool.poetry.scripts]
peac = "peac.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
gui = ["customtkinter"]

[tool.bumpversion]
current_version = "0.2.2"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
tag = true
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search = "## [Unreleased]"
replace = "## [Unreleased]\n\n### Added\n\n### Changed\n\n### Fixed\n\n## [{new_version}] - {utcnow:%Y-%m-%d}"
